Finished level design, have to beat level 7 but we're almost there. Also put the game class in a separate file to avoid bloating the main program under 200+ lines of code.
